# packages/contracts/openapi/public-api.yaml
openapi: 3.1.0
info:
  title: Smart Travel Assistant — Public API
  version: 1.0.0
paths:
...
  /api/v1/me:
    get:
      tags:
      - me
      summary: Get Me
      operationId: get_me_api_v1_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Get Me Api V1 Me Get
    patch:
      tags:
      - me
      summary: Patch Me
      operationId: patch_me_api_v1_me_patch
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MePatch'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Patch Me Api V1 Me Patch
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/me/emergency-profile:
    get:
      tags:
      - me
      summary: Get Emergency Profile
      operationId: get_emergency_profile_api_v1_me_emergency_profile_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Get Emergency Profile Api V1 Me Emergency Profile Get
    put:
      tags:
      - me
      summary: Put Emergency Profile
      operationId: put_emergency_profile_api_v1_me_emergency_profile_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmergencyProfilePut'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Put Emergency Profile Api V1 Me Emergency Profile Put
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - me
      summary: Delete Emergency Profile
      operationId: delete_emergency_profile_api_v1_me_emergency_profile_delete
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Delete Emergency Profile Api V1 Me Emergency Profile Delete
  /api/v1/consents:
    get:
      tags:
      - me
      summary: List Consents
      operationId: list_consents_api_v1_consents_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response List Consents Api V1 Consents Get
    post:
      tags:
      - me
      summary: Post Consent
      operationId: post_consent_api_v1_consents_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsentCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Post Consent Api V1 Consents Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/me/data-jobs:
    post:
      tags:
      - me
      summary: Create Data Job
      operationId: create_data_job_api_v1_me_data_jobs_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DataJobCreate'
        required: true
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Create Data Job Api V1 Me Data Jobs Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/me/data-jobs/{job_id}:
    get:
      tags:
      - me
      summary: Get Data Job
      operationId: get_data_job_api_v1_me_data_jobs__job_id__get
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Job Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Data Job Api V1 Me Data Jobs  Job Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/runs/{request_id}:
    get:
      tags:
      - runs
      summary: Get Run
      operationId: get_run_api_v1_runs__request_id__get
      parameters:
      - name: request_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Request Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Run Api V1 Runs  Request Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/runs/{request_id}/events:
    get:
      tags:
      - runs
      summary: Run Events
      operationId: run_events_api_v1_runs__request_id__events_get
      parameters:
      - name: request_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Request Id
      - name: Last-Event-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Last-Event-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/runs/{request_id}/cancel:
    post:
      tags:
      - runs
      summary: Cancel Run
      operationId: cancel_run_api_v1_runs__request_id__cancel_post
      parameters:
      - name: request_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Request Id
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - $ref: '#/components/schemas/CancelBody'
              - type: 'null'
              title: Body
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Cancel Run Api V1 Runs  Request Id  Cancel Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/runs/{request_id}/resume:
    post:
      tags:
      - runs
      summary: Resume Run
      description: 'After NEEDS_INPUT the client fixes the trip (PATCH /trips) and resumes: a new run linked to the old one.'
      operationId: resume_run_api_v1_runs__request_id__resume_post
      parameters:
      - name: request_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Request Id
      - name: Idempotency-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
            maxLength: 128
          - type: 'null'
          title: Idempotency-Key
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Resume Run Api V1 Runs  Request Id  Resume Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    CancelBody:
      properties:
        reason:
          type: string
          maxLength: 120
          title: Reason
          default: user_cancelled
      additionalProperties: false
      type: object
      title: CancelBody
    ConsentCreate:
      properties:
        type:
          $ref: '#/components/schemas/ConsentType'
        granted:
          type: boolean
          title: Granted
        policy_version:
          type: string
          maxLength: 32
          minLength: 1
          title: Policy Version
        expires_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Expires At
      additionalProperties: false
      type: object
      required:
      - type
      - granted
      - policy_version
      title: ConsentCreate
    ConsentType:
      type: string
      enum:
      - LOCATION_ONCE
      - LOCATION_LIVE
      - ALERT_NOTIFICATION
      - ANALYTICS
      - EMERGENCY_PROFILE
      title: ConsentType
...
    DataJobCreate:
      properties:
        kind:
          type: string
          enum:
          - EXPORT
          - DELETE
          title: Kind
      additionalProperties: false
      type: object
      required:
      - kind
      title: DataJobCreate
...
    EmergencyContactIn:
      properties:
        name:
          type: string
          maxLength: 120
          minLength: 1
          title: Name
        relationship:
          anyOf:
          - type: string
            maxLength: 60
          - type: 'null'
          title: Relationship
        phone:
          type: string
          maxLength: 32
          minLength: 3
          pattern: ^\+?[0-9 ()\-]{3,32}$
          title: Phone
      additionalProperties: false
      type: object
      required:
      - name
      - phone
      title: EmergencyContactIn
    EmergencyProfilePut:
      properties:
        blood_type:
          anyOf:
          - type: string
            maxLength: 8
          - type: 'null'
          title: Blood Type
        allergies:
          items:
            type: string
          type: array
          maxItems: 20
          title: Allergies
        medications:
          items:
            type: string
          type: array
          maxItems: 20
          title: Medications
        medical_notes:
          anyOf:
          - type: string
            maxLength: 2000
          - type: 'null'
          title: Medical Notes
        contacts:
          items:
            $ref: '#/components/schemas/EmergencyContactIn'
          type: array
          maxItems: 5
          title: Contacts
        insurance_provider:
          anyOf:
          - type: string
            maxLength: 120
          - type: 'null'
          title: Insurance Provider
        insurance_policy_ref:
          anyOf:
          - type: string
            maxLength: 120
          - type: 'null'
          title: Insurance Policy Ref
        insurance_phone:
          anyOf:
          - type: string
            maxLength: 32
          - type: 'null'
          title: Insurance Phone
      additionalProperties: false
      type: object
      title: EmergencyProfilePut
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    MePatch:
      properties:
        locale:
          anyOf:
          - type: string
            pattern: ^[a-z]{2}(-[A-Z]{2})?$
          - type: 'null'
          title: Locale
        timezone:
          anyOf:
          - type: string
            maxLength: 64
          - type: 'null'
          title: Timezone
        display_name:
          anyOf:
          - type: string
            maxLength: 120
          - type: 'null'
          title: Display Name
      additionalProperties: false
      type: object
      title: MePatch
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    oidc:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Keycloak realm smart-travel, audience smart-travel-api, role traveler
servers:
- url: http://localhost:8000
  description: local compose
security:
- oidc: []